save roles in db

+++ convex/files.ts
@@ -34,7 +34,8 @@ const hasAccessToOrg = async (ctx: QueryCtx | MutationCtx, orgId: string) => {
   }
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some((el) => el.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) {
     return null;
@@ -121,6 +122,16 @@ export const deleteFile = mutation({
 
     if (!access) {
       throw new ConvexError("You don't have permission to favorite this file");
+    }
+
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError(
+        "You don't have admin permission to delete this file"
+      );
     }
 
     await ctx.db.delete(args.fileId);
